+++ clang-tools-extra/clang-tidy/hooks/SlotHashBufLenCheck.h
@@ -1,0 +1,35 @@
+//===--- SlotHashBufLenCheck.h - clang-tidy ---------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_SLOTHASHBUFLENCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_SLOTHASHBUFLENCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+/// Function slot_id has fixed-size canonical hash output.
+class SlotHashBufLenCheck : public ClangTidyCheck {
+public:
+  SlotHashBufLenCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  const int64_t HASH_SIZE = 32; // documented
+  const int64_t MAX_SLOT_NO = 255; // value copied from rippled source
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_SLOTHASHBUFLENCHECK_H